+++ changelog/24404.txt
@@ -1,0 +1,3 @@
+```release-note:bug
+ui: fix issue where kv v2 capabilities checks were not passing in the full secret path if secret was inside a directory.
+```
